config/settings.py
import os
from pathlib import Path
from dotenv import load_dotenv
import dj_database_url

load_dotenv()

BASE_DIR = Path(__file__).resolve().parent.parent

SECRET_KEY = os.getenv('SECRET_KEY', 'django-insecure-change-this-key-in-production')
# Default to False for safety in production; set DEBUG=True locally via env
DEBUG = os.getenv('DEBUG', 'False') == 'True'

# Allow Railway-hosted domains and local development hosts.
# Prefer explicit `ALLOWED_HOSTS` from env; otherwise allow common local and Railway patterns.
_env_allowed = os.getenv('ALLOWED_HOSTS')
if _env_allowed:
    ALLOWED_HOSTS = [h.strip() for h in _env_allowed.split(',') if h.strip()]
else:
    ALLOWED_HOSTS = ['localhost', '127.0.0.1', '.railway.app', '.up.railway.app']

INSTALLED_APPS = [
    'django.contrib.admin',
    'django.contrib.auth',
    'django.contrib.contenttypes',
    'django.contrib.sessions',
    'django.contrib.messages',
    'django.contrib.staticfiles',
    'invoices',
]

MIDDLEWARE = [
    'django.middleware.security.SecurityMiddleware',
    'whitenoise.middleware.WhiteNoiseMiddleware',
    'django.contrib.sessions.middleware.SessionMiddleware',
    'django.middleware.common.CommonMiddleware',
    'django.middleware.csrf.CsrfViewMiddleware',
    'django.contrib.auth.middleware.AuthenticationMiddleware',
    'invoices.middleware.ActivityLogMiddleware',
    'django.contrib.messages.middleware.MessageMiddleware',
    'django.middleware.clickjacking.XFrameOptionsMiddleware',
]

# Allow same-origin framing (needed for embedded invoice previews)
X_FRAME_OPTIONS = 'SAMEORIGIN'

ROOT_URLCONF = 'config.urls'

TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        'DIRS': [BASE_DIR / 'templates'],
        'APP_DIRS': True,
        'OPTIONS': {
            'context_processors': [
                'django.template.context_processors.debug',
                'django.template.context_processors.request',
                'django.contrib.auth.context_processors.auth',
                'django.contrib.messages.context_processors.messages',
            ],
        },
    },
]

WSGI_APPLICATION = 'config.wsgi.application'

# When deployed behind a proxy (like Railway) that terminates SSL, trust the
# X-Forwarded-Proto header so Django knows the original request scheme.
SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')

# Build `CSRF_TRUSTED_ORIGINS` from env or from `ALLOWED_HOSTS`.
# If you need to override, set `CSRF_TRUSTED_ORIGINS` env to a comma-separated list
# of origins (including scheme), e.g. 'https://myapp.up.railway.app,https://example.com'.
_env_csrf = os.getenv('CSRF_TRUSTED_ORIGINS')
if _env_csrf:
    CSRF_TRUSTED_ORIGINS = [c.strip() for c in _env_csrf.split(',') if c.strip()]
else:
    CSRF_TRUSTED_ORIGINS = []
    for h in ALLOWED_HOSTS:
        if h == '*' or h == '':
            continue
        # Leading dot means allow any subdomain: '.railway.app' -> 'https://*.railway.app'
        if h.startswith('.'):
            CSRF_TRUSTED_ORIGINS.append(f'https://*{h}')
        # If hostname already contains a wildcard (e.g. '*.example.com') use it directly
        elif '*' in h:
            CSRF_TRUSTED_ORIGINS.append('https://' + h)
        else:
            # For plain hosts, add https:// prefix
            CSRF_TRUSTED_ORIGINS.append(f'https://{h}')

# Security defaults tuned per-environment
# In local development keep redirects and secure cookies off to avoid HTTPS issues
if DEBUG:
    SECURE_SSL_REDIRECT = False
    SESSION_COOKIE_SECURE = False
    CSRF_COOKIE_SECURE = False
else:
    # In production prefer secure cookies and redirect to HTTPS
    SECURE_SSL_REDIRECT = True
    SESSION_COOKIE_SECURE = True
    CSRF_COOKIE_SECURE = True

# Database
DATABASE_URL = os.getenv('DATABASE_URL')
if DATABASE_URL:
    # Use dj_database_url to parse the provided DATABASE_URL (Postgres on Railway)
    DATABASES = {
        'default': dj_database_url.config(default=DATABASE_URL, conn_max_age=600)
    }
    # Require SSL when running in production
    if not DEBUG:
        DATABASES['default'].setdefault('OPTIONS', {})
        DATABASES['default']['OPTIONS']['sslmode'] = 'require'
else:
    # Fallback to SQLite for local development
    DATABASES = {
        'default': {
            'ENGINE': 'django.db.backends.sqlite3',
            'NAME': BASE_DIR / 'db.sqlite3',
        }
    }

AUTH_PASSWORD_VALIDATORS = [
    {'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator'},
    {'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator'},
    {'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator'},
    {'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator'},
]

LANGUAGE_CODE = 'en-us'
# Set server timezone to Philippines (Manila) for display and storage
TIME_ZONE = 'Asia/Manila'
USE_I18N = True
USE_TZ = True

STATIC_URL = '/static/'
STATIC_ROOT = BASE_DIR / 'staticfiles'
STATICFILES_DIRS = [BASE_DIR / 'static']
STATICFILES_STORAGE = 'whitenoise.storage.CompressedManifestStaticFilesStorage'

MEDIA_URL = '/media/'
MEDIA_ROOT = BASE_DIR / 'media'

DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'

LOGIN_URL = 'login'
LOGIN_REDIRECT_URL = 'dashboard'
LOGOUT_REDIRECT_URL = 'login'
# Make wkhtmltopdf path configurable via env; leave unset unless using wkhtmltopdf.
WKHTMLTOPDF_CMD = os.getenv('WKHTMLTOPDF_CMD', '')

# PDF backend: default to ReportLab (pure-Python). Set `WKHTMLTOPDF_CMD` if
# you prefer wkhtmltopdf as an alternative backend.
PDF_BACKEND = os.getenv('PDF_BACKEND', 'reportlab')

# Optional: Use S3 (or S3-compatible) storage for user-uploaded media. Railway
# application filesystems can be ephemeral; enabling S3 keeps uploads persistent.
USE_S3 = os.getenv('USE_S3', 'False') in ('True', 'true', '1') or bool(os.getenv('AWS_STORAGE_BUCKET_NAME'))
if USE_S3:
    INSTALLED_APPS += ['storages']
    DEFAULT_FILE_STORAGE = 'storages.backends.s3boto3.S3Boto3Storage'
    AWS_ACCESS_KEY_ID = os.getenv('AWS_ACCESS_KEY_ID')
    AWS_SECRET_ACCESS_KEY = os.getenv('AWS_SECRET_ACCESS_KEY')
    AWS_STORAGE_BUCKET_NAME = os.getenv('AWS_STORAGE_BUCKET_NAME')
    AWS_S3_REGION_NAME = os.getenv('AWS_S3_REGION_NAME') or None
    AWS_S3_ENDPOINT_URL = os.getenv('AWS_S3_ENDPOINT_URL') or None
    AWS_QUERYSTRING_AUTH = False
    AWS_DEFAULT_ACL = os.getenv('AWS_DEFAULT_ACL', 'private')