requirements.txt
Django==5.0
psycopg2-binary==2.9.11
Pillow==9.5.0
dj-database-url==1.0.0
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0

# Pin django-storages to the latest 1.x release compatible with Python 3.11
django-storages[boto3]==1.14.6

reportlab==3.6.12